+++ packages/react-charts/src/components/ChartCursorTooltip/ChartCursorFlyout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Helpers, CommonProps, Path } from 'victory-core';
-import { isPlainObject, assign } from 'lodash';
+import isPlainObject from 'lodash/isPlainObject';
 
 const getVerticalPath = (props: any) => {
   const { pointerWidth, cornerRadius, orientation, width, height, center } = props;
@@ -89,7 +89,7 @@ const evaluateProps = (props: any) => {
   const id = Helpers.evaluateProp(props.id, props);
   const style = Helpers.evaluateStyle(props.style, props);
 
-  return assign({}, props, { id, style });
+  return Object.assign({}, props, { id, style });
 };
 
 const ChartCursorFlyout = (props: any) => {