+++ packages/react-core/src/components/Button/Button.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import styles from '@patternfly/react-styles/css/components/Button/button';
 import { css } from '@patternfly/react-styles';
+import { Spinner, spinnerSize } from '../Spinner';
 import { useOUIAProps, OUIAProps } from '../../helpers';
 
 export enum ButtonVariant {
@@ -8,6 +9,7 @@ export enum ButtonVariant {
   secondary = 'secondary',
   tertiary = 'tertiary',
   danger = 'danger',
+  warning = 'warning',
   link = 'link',
   plain = 'plain',
   control = 'control'
@@ -34,6 +36,10 @@ export interface ButtonProps extends React.HTMLProps<HTMLButtonElement>, OUIAPro
   isDisabled?: boolean;
   /** @beta Adds disabled styling and communicates that the button is disabled using the aria-disabled html attribute */
   isAriaDisabled?: boolean;
+  /** Adds progress styling to button */
+  isLoading?: boolean;
+  /** Aria-valuetext for the loading spinner */
+  spinnerAriaValueText?: string;
   /** @beta Events to prevent when the button is in an aria-disabled state */
   inoperableEvents?: string[];
   /** Adds inline styling to a link button */
@@ -41,7 +47,7 @@ export interface ButtonProps extends React.HTMLProps<HTMLButtonElement>, OUIAPro
   /** Sets button type */
   type?: 'button' | 'submit' | 'reset';
   /** Adds button variant styles */
-  variant?: 'primary' | 'secondary' | 'tertiary' | 'danger' | 'link' | 'plain' | 'control';
+  variant?: 'primary' | 'secondary' | 'tertiary' | 'danger' | 'warning' | 'link' | 'plain' | 'control';
   /** Sets position of the link icon */
   iconPosition?: 'left' | 'right';
   /** Adds accessible text to the button. */
@@ -64,6 +70,8 @@ export const Button: React.FunctionComponent<ButtonProps> = ({
   isBlock = false,
   isDisabled = false,
   isAriaDisabled = false,
+  isLoading = null,
+  spinnerAriaValueText,
   isSmall = false,
   isLarge = false,
   inoperableEvents = ['onClick', 'onKeyPress'],
@@ -121,6 +129,8 @@ export const Button: React.FunctionComponent<ButtonProps> = ({
         isAriaDisabled && styles.modifiers.ariaDisabled,
         isActive && styles.modifiers.active,
         isInline && variant === ButtonVariant.link && styles.modifiers.inline,
+        isLoading !== null && styles.modifiers.progress,
+        isLoading && styles.modifiers.inProgress,
         isSmall && styles.modifiers.small,
         isLarge && styles.modifiers.displayLg,
         className
@@ -130,6 +140,11 @@ export const Button: React.FunctionComponent<ButtonProps> = ({
       type={isButtonElement ? type : null}
       {...ouiaProps}
     >
+      {isLoading && (
+        <span className={css(styles.buttonProgress)}>
+          <Spinner size={spinnerSize.md} aria-valuetext={spinnerAriaValueText} />
+        </span>
+      )}
       {variant !== ButtonVariant.plain && icon && iconPosition === 'left' && (
         <span className={css(styles.buttonIcon, styles.modifiers.start)}>{icon}</span>
       )}