+++ packages/react-core/src/components/NotificationDrawer/NotificationDrawerList.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import { css } from '@patternfly/react-styles';
-import styles from '@patternfly/react-styles/css/components/NotificationDrawer/notification-drawer';
 
 export interface NotificationDrawerListProps extends React.HTMLProps<HTMLUListElement> {
   /**  Content rendered inside the notification drawer list body */