+++ packages/react-integration/cypress/integration/notificationdrawerbasic.spec.ts
@@ -32,6 +32,9 @@ describe('Notification Drawer Basic Demo Test', () => {
       .eq(2)
       .should('have.class', 'pf-m-hoverable');
     cy.get('.pf-c-notification-drawer__list-item')
+      .eq(3)
+      .should('have.class', 'pf-m-hoverable');
+    cy.get('.pf-c-notification-drawer__list-item')
       .last()
       .should('have.class', 'pf-m-hoverable');
   });
@@ -47,8 +50,11 @@ describe('Notification Drawer Basic Demo Test', () => {
       .eq(2)
       .should('have.class', 'pf-m-warning');
     cy.get('.pf-c-notification-drawer__list-item')
-      .last()
+      .eq(3)
       .should('have.class', 'pf-m-success');
+    cy.get('.pf-c-notification-drawer__list-item')
+      .last()
+      .should('have.class', 'pf-m-default');
   });
 
   it('Verify timestamp in list items', () => {
@@ -62,8 +68,11 @@ describe('Notification Drawer Basic Demo Test', () => {
       .eq(2)
       .contains('20 minutes ago');
     cy.get('.pf-c-notification-drawer__list-item-timestamp')
-      .last()
+      .eq(3)
       .contains('30 minutes ago');
+    cy.get('.pf-c-notification-drawer__list-item-timestamp')
+      .last()
+      .contains('35 minutes ago');
   });
 
   // Accessibility test