+++ packages/react-integration/demo-app-ts/scripts/serve.js
@@ -1,0 +1,8 @@
+const LocalWebServer = require('local-web-server');
+
+const localWebServer = new LocalWebServer();
+localWebServer.listen({
+  directory: './public',
+  port: 3000,
+  spa: 'index.html'
+});