+++ packages/react-integration/demo-app-ts/src/components/demos/DonutChartDemo/DonutRightAlignedLegendColorDemo.tsx
@@ -12,7 +12,11 @@ export class DonutRightAlignedLegendColorDemo extends React.Component {
       <div>
         <div style={{ height: '230px', width: '350px', paddingTop: '50px', paddingLeft: '50px' }}>
           <ChartDonut
-            data={[{ x: 'Cats', y: 35 }, { x: 'Dogs', y: 55 }, { x: 'Birds', y: 10 }]}
+            data={[
+              { x: 'Cats', y: 35 },
+              { x: 'Dogs', y: 55 },
+              { x: 'Birds', y: 10 }
+            ]}
             labels={({ datum }) => `${datum.x}: ${datum.y}%`}
             legendData={[{ name: 'Cats: 35' }, { name: 'Dogs: 55' }, { name: 'Birds: 10' }]}
             legendOrientation="vertical"