+++ packages/react-integration/demo-app-ts/src/components/demos/PieChartDemo/PieColorDemo.tsx
@@ -12,7 +12,11 @@ export class PieColorDemo extends React.Component {
       <div style={{ height: '275px', width: '300px', paddingTop: '50px', paddingLeft: '50px' }}>
         <div>
           <ChartPie
-            data={[{ x: 'Cats', y: 35 }, { x: 'Dogs', y: 55 }, { x: 'Birds', y: 10 }]}
+            data={[
+              { x: 'Cats', y: 35 },
+              { x: 'Dogs', y: 55 },
+              { x: 'Birds', y: 10 }
+            ]}
             height={275}
             labels={({ datum }) => `${datum.x}: ${datum.y}`}
             legendData={[{ name: 'Cats: 35' }, { name: 'Dogs: 55' }, { name: 'Birds: 10' }]}