+++ packages/react-integration/demo-app-ts/src/components/demos/TabsDemo/TabsDemo.tsx
@@ -168,6 +168,23 @@ export class TabDemo extends Component {
             Tab 3 section
           </Tab>
         </Tabs>
+        <Tabs
+          id="boxLightVariantTabs"
+          activeKey={this.state.activeTabKey4}
+          onSelect={this.handleTabClick4}
+          isBox
+          variant="light300"
+        >
+          <Tab eventKey={0} title={<TabTitleText>Tab item 1</TabTitleText>}>
+            Tab 1 section
+          </Tab>
+          <Tab eventKey={1} title={<TabTitleText>Tab item 2</TabTitleText>}>
+            Tab 2 section
+          </Tab>
+          <Tab eventKey={2} title={<TabTitleText>Tab item </TabTitleText>}>
+            Tab 3 section
+          </Tab>
+        </Tabs>
       </React.Fragment>
     );
   }