+++ packages/react-table/src/components/Table/utils/decorators/classNames.ts
@@ -32,14 +32,11 @@ interface Visibility {
 
 export const Visibility = visibilityModifiers
   .filter(key => styles.modifiers[key])
-  .reduce(
-    (acc, curr) => {
-      const key2 = curr.replace('2xl', '2Xl') as keyof typeof Visibility;
-      acc[key2] = styles.modifiers[curr];
-      return acc;
-    },
-    {} as Visibility
-  );
+  .reduce((acc, curr) => {
+    const key2 = curr.replace('2xl', '2Xl') as keyof typeof Visibility;
+    acc[key2] = styles.modifiers[curr];
+    return acc;
+  }, {} as Visibility);
 
 export const classNames = (...classes: string[]): ITransform => () => ({
   className: css(...classes)