+++ packages/react-catalog-view-extension/CHANGELOG.md
@@ -3,6 +3,105 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.42](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.41...@patternfly/react-catalog-view-extension@4.8.42) (2020-09-25)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.41](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.40...@patternfly/react-catalog-view-extension@4.8.41) (2020-09-25)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.40](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.39...@patternfly/react-catalog-view-extension@4.8.40) (2020-09-24)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.39](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.38...@patternfly/react-catalog-view-extension@4.8.39) (2020-09-24)
+
+
+### Bug Fixes
+
+* **demo-app:** replace CRA with webpack and ts-loader ([#4877](https://github.com/patternfly/patternfly-react/issues/4877)) ([1b64309](https://github.com/patternfly/patternfly-react/commit/1b6430943e14acf84fa54115b1a009930cf2372e))
+
+
+
+
+
+## [4.8.38](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.37...@patternfly/react-catalog-view-extension@4.8.38) (2020-09-23)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.37](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.36...@patternfly/react-catalog-view-extension@4.8.37) (2020-09-23)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.36](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.35...@patternfly/react-catalog-view-extension@4.8.36) (2020-09-23)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.35](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.34...@patternfly/react-catalog-view-extension@4.8.35) (2020-09-22)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.34](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.33...@patternfly/react-catalog-view-extension@4.8.34) (2020-09-21)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.32...@patternfly/react-catalog-view-extension@4.8.33) (2020-09-21)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.31...@patternfly/react-catalog-view-extension@4.8.32) (2020-09-21)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.30...@patternfly/react-catalog-view-extension@4.8.31) (2020-09-18)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
 ## [4.8.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.29...@patternfly/react-catalog-view-extension@4.8.30) (2020-09-17)
 
 **Note:** Version bump only for package @patternfly/react-catalog-view-extension